classic_sim/requirements.txt
atomicwrites==1.4.1
cma==4.0.0
colorhash==2.0.0
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.55.8
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tqdm==4.67.1
